Added: Program of the Multiple Rides

+++ src/test/java/com/cabinvoicegenerator/CabInvoiceGeneratorTest.java
@@ -9,7 +9,7 @@ public class CabInvoiceGeneratorTest {
     @Test
     public void testCabInvoiceGenerator() { //step1: This  is my test case for Calculate Fare
         double distance = 2.0;
-        int time = 5;
+        double time = 5.0;
         double result = cabInvoiceGenerator.calculateFare(distance, time);
         Assert.assertEquals(25, result,0.0);
     }
